// src/lib/api/__generated/openapi.ts/zod.gen.ts
export const zBisItem = z.object({
    itemId: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }),
    itemName: z.string(),
    slot: z.string(),
    priority: z.int(),
    source: z.string().nullish(),
    note: z.string().nullish()
});
// ...
export const zSaveBisListRequest = z.object({
    expectedVersion: z.int().nullish(),
    phaseOrTier: z.string().nullish(),
    items: z.array(zBisItem)
});
// ...
export const zSaveBisListBody = zSaveBisListRequest;